Refactor mutation operators for correctness, quality, and readability
- Created `MutationUtils` to centralize common logic (`getDefault`, `getLocal`, `getName`, `getNextIndex`, `getDistance`, `cast`, `copy`).
- Refactored `ReplaceBitwiseOperator` to be stateless (thread-safe) and use `MutationUtils`.
- Refactored `InsertUnaryOperator` to use stack manipulation instead of `IINC` to avoid side effects on local variables, and used `MutationUtils`.
- Refactored `ReplaceVariable` to use `MutationUtils` and remove duplicated code.
- Refactored `DeleteStatement` and `DeleteField` to use `MutationUtils` and improve `INVOKEINTERFACE` handling (removed unsafe reflection).
- Refactored `ReplaceArithmeticOperator` to use `MutationUtils`.
- Refactored `ReplaceComparisonOperator` to reduce massive code duplication in `getInfectionDistance` by canonicalizing opcodes.
- Removed unused imports and cleaned up code.
