+++ client/src/main/java/org/evosuite/instrumentation/mutation/DeleteStatement.java
@@ -31,7 +31,6 @@ import org.objectweb.asm.tree.analysis.Frame;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.lang.reflect.Modifier;
 import java.util.LinkedList;
 import java.util.List;
 
@@ -81,27 +80,10 @@ public class DeleteStatement implements MutationOperator {
             logger.debug("Deleting callee of type " + node.owner);
             mutation.add(new InsnNode(Opcodes.POP));
         } else if (node.getOpcode() == Opcodes.INVOKEINTERFACE) {
-            boolean isStatic = false;
-            try {
-                Class<?> clazz = Class.forName(node.owner.replace('/', '.'), false, TestGenerationContext.getInstance().getClassLoaderForSUT());
-                for (java.lang.reflect.Method method : clazz.getMethods()) {
-                    if (method.getName().equals(node.name)) {
-                        if (Type.getMethodDescriptor(method).equals(node.desc)) {
-                            if (Modifier.isStatic(method.getModifiers()))
-                                isStatic = true;
-                        }
-                    }
-                }
-            } catch (ClassNotFoundException e) {
-                logger.warn("Could not find class: " + node.owner
-                        + ", this is likely a severe problem");
-            }
-            if (!isStatic) {
-                logger.debug("Deleting callee of type " + node.owner);
-                mutation.add(new InsnNode(Opcodes.POP));
-            }
+            logger.debug("Deleting callee of type " + node.owner);
+            mutation.add(new InsnNode(Opcodes.POP));
         }
-        mutation.add(getDefault(returnType));
+        mutation.add(MutationUtils.getDefault(returnType));
 
         // insert mutation into pool
         Mutation mutationObject = MutationPool.getInstance(TestGenerationContext.getInstance().getClassLoaderForSUT()).addMutation(className,
@@ -115,30 +97,6 @@ public class DeleteStatement implements MutationOperator {
 
         mutations.add(mutationObject);
         return mutations;
-    }
-
-    private static AbstractInsnNode getDefault(Type type) {
-        if (type.equals(Type.BOOLEAN_TYPE)) {
-            return new LdcInsnNode(0);
-        } else if (type.equals(Type.INT_TYPE)) {
-            return new LdcInsnNode(0);
-        } else if (type.equals(Type.BYTE_TYPE)) {
-            return new LdcInsnNode(0);
-        } else if (type.equals(Type.CHAR_TYPE)) {
-            return new LdcInsnNode(0);
-        } else if (type.equals(Type.DOUBLE_TYPE)) {
-            return new LdcInsnNode(0.0);
-        } else if (type.equals(Type.FLOAT_TYPE)) {
-            return new LdcInsnNode(0.0F);
-        } else if (type.equals(Type.LONG_TYPE)) {
-            return new LdcInsnNode(0L);
-        } else if (type.equals(Type.SHORT_TYPE)) {
-            return new LdcInsnNode(0);
-        } else if (type.equals(Type.VOID_TYPE)) {
-            return new LabelNode();
-        } else {
-            return new InsnNode(Opcodes.ACONST_NULL);
-        }
     }
 
     /* (non-Javadoc)
